chore: solana adapter improvements
