+++ examples/deposit-flow/package.json
@@ -13,21 +13,21 @@
     "@lifi/sdk": "^3.6.7",
     "@lifi/widget": "workspace:^",
     "@mui/material": "^6.4.10",
-    "@tanstack/react-query": "^5.72.0",
+    "@tanstack/react-query": "^5.74.2",
     "events": "^3.3.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "wagmi": "^2.14.16"
   },
   "devDependencies": {
     "@types/events": "^3.0.3",
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react": "^4.3.4",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vite-plugin-node-polyfills": "^0.23.0"
   }
 }
