+++ examples/dynamic/package.json
@@ -9,13 +9,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@dynamic-labs/bitcoin": "^4.10.4",
-    "@dynamic-labs/ethereum": "^4.10.4",
-    "@dynamic-labs/ethereum-aa": "^4.10.4",
-    "@dynamic-labs/sdk-react-core": "^4.10.4",
-    "@dynamic-labs/solana": "^4.10.4",
-    "@dynamic-labs/solana-core": "^4.10.4",
-    "@dynamic-labs/wagmi-connector": "^4.10.4",
+    "@dynamic-labs/bitcoin": "^4.11.2",
+    "@dynamic-labs/ethereum": "^4.11.2",
+    "@dynamic-labs/ethereum-aa": "^4.11.2",
+    "@dynamic-labs/sdk-react-core": "^4.11.2",
+    "@dynamic-labs/solana": "^4.11.2",
+    "@dynamic-labs/solana-core": "^4.11.2",
+    "@dynamic-labs/wagmi-connector": "^4.11.2",
     "@lifi/wallet-management": "workspace:^",
     "@lifi/widget": "workspace:^",
     "@mui/material": "^6.4.10",
@@ -23,7 +23,7 @@
     "@solana/wallet-adapter-react": "^0.15.36",
     "@solana/wallet-standard-features": "^1.3.0",
     "@solana/web3.js": "^1.98.0",
-    "@tanstack/react-query": "^5.72.0",
+    "@tanstack/react-query": "^5.74.2",
     "@wagmi/core": "^2.16.4",
     "@wallet-standard/base": "^1.1.0",
     "@wallet-standard/features": "^1.1.0",
@@ -31,16 +31,16 @@
     "mitt": "^3.0.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "vite-plugin-env-compatible": "^2.0.1",
     "wagmi": "^2.14.16"
   },
   "devDependencies": {
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react-swc": "^3.8.1",
     "globals": "^16.0.0",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5"
+    "vite": "^6.2.6"
   }
 }
