+++ examples/nextjs-page-router/package.json
@@ -8,15 +8,15 @@
     "start": "next start"
   },
   "dependencies": {
-    "@lifi/widget": "^3.18.2",
-    "next": "^15.2.4",
+    "@lifi/widget": "^3.18.4",
+    "next": "^15.3.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "typescript": "^5.8.3"
   }
 }
