+++ examples/nextjs/package.json
@@ -9,16 +9,16 @@
   },
   "dependencies": {
     "@lifi/sdk": "^3.6.7",
-    "@lifi/widget": "^3.18.2",
+    "@lifi/widget": "^3.18.4",
     "@mui/material-nextjs": "^6.4.3",
-    "next": "^15.2.4",
+    "next": "^15.3.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "typescript": "^5.8.3"
   }
 }
