+++ examples/privy/package.json
@@ -10,30 +10,30 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@lifi/wallet-management": "^3.7.2",
-    "@lifi/widget": "^3.18.2",
+    "@lifi/wallet-management": "^3.7.3",
+    "@lifi/widget": "^3.18.4",
     "@mui/icons-material": "6.0.2",
     "@mui/material": "^6.4.10",
-    "@privy-io/react-auth": "^2.8.2",
+    "@privy-io/react-auth": "^2.8.3",
     "@privy-io/wagmi": "^1.0.3",
     "@solana/wallet-adapter-base": "^0.9.24",
     "@solana/wallet-adapter-react": "^0.15.36",
     "@solana/web3.js": "^1.98.0",
-    "@tanstack/react-query": "^5.72.0",
+    "@tanstack/react-query": "^5.74.2",
     "mitt": "^3.0.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "wagmi": "^2.14.16"
   },
   "devDependencies": {
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react": "^4.3.4",
     "globals": "^16.0.0",
     "typescript": "~5.8.3",
-    "typescript-eslint": "^8.29.0",
-    "vite": "^6.2.5",
+    "typescript-eslint": "^8.29.1",
+    "vite": "^6.2.6",
     "vite-plugin-node-polyfills": "^0.23.0"
   }
 }
