+++ examples/remix/package.json
@@ -10,22 +10,22 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@lifi/widget": "^3.18.2",
-    "@remix-run/css-bundle": "^2.16.4",
-    "@remix-run/node": "^2.16.4",
-    "@remix-run/react": "^2.16.4",
-    "@remix-run/serve": "^2.16.4",
-    "isbot": "^5.1.25",
+    "@lifi/widget": "^3.18.4",
+    "@remix-run/css-bundle": "^2.16.5",
+    "@remix-run/node": "^2.16.5",
+    "@remix-run/react": "^2.16.5",
+    "@remix-run/serve": "^2.16.5",
+    "isbot": "^5.1.26",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "remix-utils": "^8.5.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "^2.16.4",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@remix-run/dev": "^2.16.5",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vite-tsconfig-paths": "^5.1.4"
   },
   "engines": {
