+++ examples/svelte/package.json
@@ -13,15 +13,15 @@
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@tsconfig/svelte": "^5.0.4",
     "@types/events": "^3.0.3",
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
-    "svelte": "^5.25.7",
-    "svelte-check": "^4.1.5",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
+    "svelte": "^5.26.2",
+    "svelte-check": "^4.1.6",
     "svelte-preprocess": "^6.0.3",
     "tslib": "^2.8.1",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vite-plugin-node-polyfills": "^0.23.0"
   },
   "dependencies": {
