+++ examples/vue/package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@lifi/widget": "^3.18.2",
+    "@lifi/widget": "^3.18.4",
     "veaury": "^2.6.2",
     "vue": "^3.5.13"
   },
@@ -18,7 +18,7 @@
     "@vitejs/plugin-vue": "^5.2.3",
     "@vitejs/plugin-vue-jsx": "^4.1.2",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vite-plugin-node-polyfills": "^0.23.0",
     "vue-tsc": "^2.2.8"
   }
