+++ package.json
@@ -49,14 +49,14 @@
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
     "@types/events": "^3.0.3",
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "cpy-cli": "^5.0.0",
     "fs-extra": "^11.3.0",
     "husky": "^9.1.7",
-    "lerna": "8.2.1",
-    "lint-staged": "^15.5.0",
+    "lerna": "8.2.2",
+    "lint-staged": "^15.5.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "standard-version": "^9.5.0",
@@ -65,11 +65,8 @@
   "pnpm": {
     "overrides": {
       "bs58": ">=4.0.1",
-      "@tanstack/react-query": ">=5.67.3",
-      "@solana/wallet-adapter-base": ">=0.9.24",
-      "wagmi": "2.14.0",
       "@safe-global/safe-apps-provider": ">=0.18.6"
     }
   },
-  "packageManager": "pnpm@10.7.1+sha512.2d92c86b7928dc8284f53494fb4201f983da65f0fb4f0d40baafa5cf628fa31dae3e5968f12466f17df7e97310e30f343a648baea1b9b350685dafafffdf5808"
+  "packageManager": "pnpm@10.8.1+sha512.c50088ba998c67b8ca8c99df8a5e02fd2ae2e2b29aaf238feaa9e124248d3f48f9fb6db2424949ff901cffbb5e0f0cc1ad6aedb602cd29450751d11c35023677"
 }
