+++ packages/wallet-management/package.json
@@ -62,7 +62,7 @@
     "mitt": "^3.0.1",
     "react-i18next": "^15.4.1",
     "use-sync-external-store": "^1.5.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
