+++ packages/widget-embedded/package.json
@@ -20,14 +20,14 @@
     "@mui/material": "^6.4.10",
     "@mui/system": "^6.4.10",
     "@opensea/seaport-js": "4.0.4",
-    "@tanstack/react-query": "^5.72.0",
-    "bignumber.js": "^9.2.0",
+    "@tanstack/react-query": "^5.74.2",
+    "bignumber.js": "^9.2.1",
     "ethers": "^6.13.5",
     "events": "^3.3.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router-dom": "^6.30.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "wagmi": "^2.14.16"
   },
   "devDependencies": {
@@ -35,7 +35,7 @@
     "@vitejs/plugin-react-swc": "^3.8.1",
     "source-map-explorer": "^2.5.3",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vite-plugin-node-polyfills": "^0.23.0",
     "web-vitals": "^4.2.4"
   },
