+++ packages/widget-playground-next/package.json
@@ -16,16 +16,16 @@
     "@mui/material": "^6.4.10",
     "@mui/material-nextjs": "^6.4.3",
     "@mui/system": "^6.4.10",
-    "@tanstack/react-query": "^5.72.0",
+    "@tanstack/react-query": "^5.74.2",
     "core-js": "^3.41.0",
-    "next": "^15.2.4",
+    "next": "^15.3.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "typescript": "^5.8.3"
   },
   "private": true
