+++ packages/widget-playground-vite/package.json
@@ -18,7 +18,7 @@
     "@lifi/widget-playground": "workspace:*",
     "@mui/material": "^6.4.10",
     "@mui/system": "^6.4.10",
-    "@tanstack/react-query": "^5.72.0",
+    "@tanstack/react-query": "^5.74.2",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-router-dom": "^6.30.0"
@@ -29,7 +29,7 @@
     "rollup-plugin-polyfill-node": "^0.13.0",
     "source-map-explorer": "^2.5.3",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vite-plugin-node-polyfills": "0.23.0",
     "web-vitals": "^4.2.4"
   },
