+++ packages/widget-playground/package.json
@@ -35,17 +35,16 @@
     "@lifi/widget": "workspace:*",
     "@monaco-editor/react": "^4.7.0",
     "@mui/icons-material": "6.0.2",
+    "@mui/lab": "6.0.0-beta.31",
+    "@mui/material": "^6.4.10",
+    "@mui/system": "^6.4.10",
+    "@rainbow-me/rainbowkit": "^2.2.4",
     "@reown/appkit": "^1.7.0",
     "@reown/appkit-adapter-bitcoin": "^1.7.0",
     "@reown/appkit-adapter-solana": "^1.7.0",
     "@reown/appkit-adapter-wagmi": "^1.7.0",
     "@reown/appkit-common": "^1.7.0",
     "@reown/appkit-utils": "^1.7.0",
-    "@solana/wallet-adapter-wallets": "^0.19.32",
-    "@mui/lab": "6.0.0-beta.31",
-    "@mui/material": "^6.4.10",
-    "@mui/system": "^6.4.10",
-    "@rainbow-me/rainbowkit": "^2.2.4",
     "@solana/wallet-adapter-base": "^0.9.24",
     "@solana/wallet-adapter-react": "^0.15.36",
     "@solana/web3.js": "^1.98.0",
@@ -55,20 +54,20 @@
     "mitt": "^3.0.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "wagmi": "^2.14.16",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
     "@types/lodash.isequal": "^4.5.8",
-    "@types/node": "^22.14.0",
-    "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/node": "^22.14.1",
+    "@types/react": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react-swc": "^3.8.1",
     "cpy-cli": "^5.0.0",
     "madge": "^8.0.0",
     "typescript": "^5.8.3",
-    "vite": "^6.2.5",
+    "vite": "^6.2.6",
     "vitest": "^3.1.1"
   },
   "peerDependencies": {
