+++ packages/widget/package.json
@@ -67,7 +67,7 @@
     "react-i18next": "^15.4.1",
     "react-intersection-observer": "^9.16.0",
     "react-router-dom": "^6.30.0",
-    "viem": "^2.26.0",
+    "viem": "^2.26.5",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
